Consolidate constants in Settings and make Motor constants
