+++ src/main/java/com/stuypulse/robot/constants/Motors.java
@@ -1,0 +1,9 @@
+package com.stuypulse.robot.constants;
+
+import com.revrobotics.CANSparkMax.IdleMode;
+import com.stuypulse.robot.util.SparkMaxConfig;
+
+public interface Motors {
+	SparkMaxConfig TURN_CONFIG = new SparkMaxConfig(false, IdleMode.kBrake, 60, 0.0);
+	SparkMaxConfig DRIVE_CONFIG = new SparkMaxConfig(false, IdleMode.kBrake, 60, 0.0);
+}
