+++ src/main/java/com/stuypulse/robot/subsystems/modules/SL_SwerveModule.java
@@ -2,10 +2,10 @@ package com.stuypulse.robot.subsystems.modules;
 
 import com.revrobotics.CANSparkMax;
 import com.revrobotics.RelativeEncoder;
-import com.revrobotics.CANSparkMax.IdleMode;
 import com.revrobotics.CANSparkMaxLowLevel.MotorType;
+import com.stuypulse.robot.constants.Motors;
+import com.stuypulse.robot.constants.Settings.Robot.Encoder;
 import com.stuypulse.robot.subsystems.SwerveModule;
-import com.stuypulse.robot.util.SparkMaxConfig;
 import com.stuypulse.stuylib.control.Controller;
 import com.stuypulse.stuylib.control.angle.AngleController;
 import com.stuypulse.stuylib.control.angle.feedback.AnglePIDController;
@@ -19,7 +19,6 @@ import edu.wpi.first.math.MathUtil;
 import edu.wpi.first.math.geometry.Rotation2d;
 import edu.wpi.first.math.geometry.Translation2d;
 import edu.wpi.first.math.kinematics.SwerveModuleState;
-import edu.wpi.first.math.util.Units;
 import edu.wpi.first.wpilibj.DutyCycleEncoder;
 import edu.wpi.first.wpilibj.smartdashboard.SmartDashboard;
 import edu.wpi.first.wpilibj2.command.SubsystemBase;
@@ -40,26 +39,6 @@ public class SL_SwerveModule extends SubsystemBase implements SwerveModule {
         SmartNumber kS = new SmartNumber("Swerve/Drive/kS", 0.11114);
         SmartNumber kV = new SmartNumber("Swerve/Drive/kV", 2.7851);
         SmartNumber kA = new SmartNumber("Swerve/Drive/kA", 0.30103);
-    }
-
-    public interface Encoder {
-        public interface Stages {
-            // input / output
-            double FIRST = 16.0 / 48.0;
-            double SECOND = 28.0 / 16.0;
-            double THIRD = 15.0 / 60.0;
-        }
-
-        double WHEEL_CIRCUMFERENCE = Math.PI * Units.inchesToMeters(4.0);
-        double GEAR_RATIO = Stages.FIRST * Stages.SECOND * Stages.THIRD;
-
-        double POSITION_CONVERSION = WHEEL_CIRCUMFERENCE * GEAR_RATIO;
-        double VELOCITY_CONVERSION = POSITION_CONVERSION / 60.0;
-    }
-
-    public interface Motors {
-        SparkMaxConfig Turn = new SparkMaxConfig(false, IdleMode.kBrake, 40, 0.0);
-        SparkMaxConfig Drive = new SparkMaxConfig(false, IdleMode.kBrake, 60, 0.0);
     }
 
     // module data
@@ -95,7 +74,7 @@ public class SL_SwerveModule extends SubsystemBase implements SwerveModule {
         // turn
 
         turnMotor = new CANSparkMax(turnCANId, MotorType.kBrushless);
-        Motors.Turn.config(turnMotor);
+        Motors.TURN_CONFIG.config(turnMotor);
 
         absoluteEncoder = new DutyCycleEncoder(absoluteEncoderChannel);
         this.angleOffset = angleOffset;
@@ -104,10 +83,10 @@ public class SL_SwerveModule extends SubsystemBase implements SwerveModule {
         // drive
 
         driveMotor = new CANSparkMax(driveCANId, MotorType.kBrushless);
-        Motors.Drive.config(driveMotor);
+        Motors.DRIVE_CONFIG.config(driveMotor);
         driveEncoder = driveMotor.getEncoder();
-        driveEncoder.setPositionConversionFactor(Encoder.POSITION_CONVERSION);
-        driveEncoder.setVelocityConversionFactor(Encoder.VELOCITY_CONVERSION);
+        driveEncoder.setPositionConversionFactor(Encoder.Drive.POSITION_CONVERSION);
+        driveEncoder.setVelocityConversionFactor(Encoder.Drive.VELOCITY_CONVERSION);
 
         driveController = new PIDController(Drive.kP, Drive.kI, Drive.kD)
                 .add(new Feedforward.Motor(Drive.kS, Drive.kV, Drive.kA).velocity());
